Another signed/unsigned warning

+++ src/leveldb/util/logging.cc
@@ -64,7 +64,7 @@ bool ConsumeDecimalNumber(Slice* in, uint64_t* val) {
       const int delta = (c - '0');
       static const uint64_t kMaxUint64_t = ~static_cast<uint64_t>(0);
       if (v > kMaxUint64_t/10 ||
-          (v == kMaxUint64_t/10 && delta > kMaxUint64_t%10)) {
+          (v == kMaxUint64_t/10 && delta > (int)(kMaxUint64_t%10))) {
         // Overflow
         return false;
       }
